% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\title{create db connection}
\usage{
db_connect()
}
\value{
conSuper returning connection for use in other functions
}
\description{
create db connection
}
\examples{
db_connect()
}
