% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_merge.R
\name{out_merge}
\alias{out_merge}
\alias{rbind_diff_col}
\title{Outer appends files with different columns}
\usage{
rbind_diff_col(x, y)
}
\arguments{
\item{x}{first dataset}

\item{y}{second dataset}
}
\description{
Outer appends files with different columns
}
\examples{
df1 = data.frame(CustomerId = c(1:6), Product = c(rep("Toaster", 3), rep("Radio", 3)))
df2 = data.frame(CustomerId = c(2, 4, 6), State = c(rep("Alabama", 2), rep("Ohio", 1)))
rbind_diff_col(df1,df2)
}
