% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_ld.R
\name{scan_ld}
\alias{scan_ld}
\title{Scan London Data store for resources}
\arguments{
\item{test}{1 if simulate update event, 0 otherwise}
}
\value{
df_dplyr dataset for analysis
}
\description{
Scan London Data store for resources
}
\examples{
scan_ld()
}
