% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_changedate.R
\name{test_date}
\alias{test_date}
\title{simulate update for testing}
\usage{
test_date(dataset = ldndata)
}
\arguments{
\item{dataset}{dataset to simulate update for testing}
}
\description{
simulate update for testing
}
\examples{
dataset <- data.frame(matrix(ncol = 7, nrow = 0))
colnames(dataset) <- c("ido", "id", "orderupdo", "uptdwheno", "check_timestampo", "check_timestamp", "slugo")
test_date(dataset)
}
