% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_exists.R
\name{url_exists}
\alias{url_exists}
\title{Checks whether url is working}
\usage{
url_exists(x, non_2xx_return_value = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{x}{url to check}

\item{non_2xx_return_value}{value to return when url is false}

\item{quiet}{whether to suppress warnings}

\item{...}{other arguments}
}
\description{
Checks whether url is working
}
\examples{
url_exists("https://google.com")
}
